new footer test
this patch also has some new helper function to make the code more efficient.

one of the function is extract_link that accepts links from a perticular HTML tag. then it
scrapes all the links in that html tag.

this is the core for the footer test and it is error and bug free at this point.